refactor: cache helper
