+++ src/api/fetchBase64Image.ts
@@ -1,7 +1,7 @@
 import { pipe } from 'fp-ts/lib/function';
 import * as TE from 'fp-ts/lib/TaskEither';
 import { GatsbyCache } from 'gatsby';
-import { withCache } from '../modules/gatsby-source-url/cache';
+import { withCache } from '../common/cache';
 import { fetch } from '../modules/gatsby-source-url/utils';
 
 export const buildBase64URL = (contentType: string, base64: string): string =>
